Refactor vision operators into separate packages
