+++ src/Aeon.Acquisition/Aeon.Acquisition.csproj
@@ -17,22 +17,14 @@
     <PackageReference Include="Bonsai.Core" Version="2.8.1" />
     <PackageReference Include="Bonsai.Audio" Version="2.8.0" />
     <PackageReference Include="Bonsai.Harp" Version="3.5.0" />
-    <PackageReference Include="Harp.Behavior" Version="0.1.0" />
-    <PackageReference Include="Harp.CameraControllerGen2" Version="0.1.0" />
     <PackageReference Include="Harp.ClockSynchronizer" Version="0.1.0" />
-    <PackageReference Include="Harp.OutputExpander" Version="0.2.0" />
     <PackageReference Include="Harp.TimestampGeneratorGen3" Version="0.1.0" />
-    <PackageReference Include="Bonsai.Osc" Version="2.7.0" />
-    <PackageReference Include="Bonsai.Pylon" Version="0.3.0" />
-    <PackageReference Include="Bonsai.Sleap" Version="0.3.0" />
-    <PackageReference Include="Bonsai.Spinnaker" Version="0.7.1" />
     <PackageReference Include="Bonsai.System" Version="2.8.0" />
     <PackageReference Include="Bonsai.Scripting.Expressions" Version="2.8.0" />
-    <PackageReference Include="Bonsai.ZeroMQ" Version="0.2.0" />
     <PackageReference Include="LibGit2Sharp" Version="0.25.4" />
-    <PackageReference Include="Bonsai.Vision" Version="2.8.0" />
     <PackageReference Include="Bonsai.Numerics" Version="0.9.0" />
     <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
+    <PackageReference Include="YamlDotNet" Version="13.1.1" />
   </ItemGroup>
 
 </Project>
