+++ src/Aeon.Environment/Aeon.Environment.csproj
@@ -6,7 +6,7 @@
     <Description>Provides reactive modules for controlling and monitoring Project Aeon environments.</Description>
     <PackageTags>Bonsai Rx Project Aeon Environment</PackageTags>
     <TargetFramework>net472</TargetFramework>
-    <VersionPrefix>0.1.0</VersionPrefix>
+    <VersionPrefix>0.2.0</VersionPrefix>
     <VersionSuffix></VersionSuffix>
   </PropertyGroup>
   
@@ -16,6 +16,8 @@
 
   <ItemGroup>
     <PackageReference Include="Bonsai.Design" Version="2.8.0" />
+    <PackageReference Include="Bonsai.Osc" Version="2.7.0" />
+    <PackageReference Include="Bonsai.ZeroMQ" Version="0.2.0" />
     <PackageReference Include="Harp.RfidReader" Version="0.1.0" />
   </ItemGroup>
