+++ src/Aeon.Tests/Aeon.Tests.csproj
@@ -24,6 +24,9 @@
     <ProjectReference Include="..\Aeon.Acquisition\Aeon.Acquisition.csproj" />
     <ProjectReference Include="..\Aeon.Environment\Aeon.Environment.csproj" />
     <ProjectReference Include="..\Aeon.Foraging\Aeon.Foraging.csproj" />
+    <ProjectReference Include="..\Aeon.Video\Aeon.Video.csproj" />
+    <ProjectReference Include="..\Aeon.Vision.Sleap\Aeon.Vision.Sleap.csproj" />
+    <ProjectReference Include="..\Aeon.Vision\Aeon.Vision.csproj" />
   </ItemGroup>
 
 </Project>
