+++ src/Aeon.Video/CameraController.bonsai
