+++ src/Aeon.Video/Properties/AssemblyInfo.cs
@@ -1,0 +1,6 @@
+﻿using Bonsai;
+
+// General Information about an assembly is controlled through the following 
+// set of attributes. Change these attribute values to modify the information
+// associated with an assembly.
+[assembly: XmlNamespacePrefix("clr-namespace:Aeon.Video", "aeon-video")]
